Fixed the minimization bug
The bug with minimization was fixed, and the issue was that I had named
the list with energies that I collected "energy", but there is a jaxmd
function called energy that is called for the simple_spring_bond
function, and so it thinks energy is a list and not a function. Easy
solution was to rename it. This didn't happen in first minimization as
I had called that list E_init.