Implemented harmonic walls that act only on edge nodes
The harmonic walls now only act on the left and right edge nodes.
There is still some weird behavior in that the walls don't seem to move
much if I compress it.

Also of course, I changed the sign of the harmonic walls, since we want
them to be repulsive and decrease in energy as edge nodes move away.