Split up minimize functions
Modifying the code to work with jit. This is a work in progress.
The new minimize functions work without jit, but not with it.
I may need to get rid of the for loops and change the control flow
within the minimize functions. I will try to do this later.