Solved force issues for second minimization
Pablo helped me solve the issues. The issue was that the fire_apply
argument was only read once at compile time, so it was not updated
for the second minimization, and so in the second minimization it used
the first minimization's fire_apply. The solution was to make the
fire_apply an argument to a new function, and then call that function.