Simplified log output and poisson calculations
I made the log output such that a single log variable is used for both
first and second minimization, and could easily be used to
visualize both minimization in the same visualization. I also changed
the poisson calculation in the main function to just use the non-corner
nodes for calculating the poisson ratio.