Made code modular and added a main file
I split the functions in auxetic.ipynb into separate python files
and added a main file to run the simulation. I also modified the
angular calculation function as there was some problem with it after
modularizing the code. The modulaized code works with simulate_auxetic
function, but I have not tested it with other optimizing functions yet.