Made code run optimize_aging_compression as well.
In addition to the changes in the previous commit, I also tested the
function about creating acoustic bandgaps to make sure they are working.
I had to make some modifications to the simulation.py file where I just
made sure that some arguments are passed explicitly.

I noticed that the code is very cumbersome, with some functions having
20 arguments. To make the code more readable, it might be good to
introduce a class that holds all the parameters and then pass that to
the functions.