Improved code of acoustic functions
I continued to clean up code and make the acoustic functions easier
to understand. I still need to continue cleaning it a bit, and
add some more properties to the class System, so that I can have
even fewer arguments in the functions. Then I need to test it all.