Fixed bug in angle calculations
I Fixed a bug in the angle calculations. The angle was calculated by
accidentally vectorizing the angle energy twice. The initial angles were
already vectorized, and then I calculated current angles angle by angle
and then vectorized it, so that each angle took minus ALL of the initial
system angles instead of just the one. This is now fixed.