Fixed the issue of forces being nans
The issue was in the angle calculation, where np.clip was used to make
sure the cos(theta) was between -1 and 1. However, this probably
produced some instability in JAX MD taking gradients wrt to angles when
calculating the forces through quantity.force(energy_fn). Because
the problematic angle was exactly pi, so what I needed to do was to
change the np.clip to go from -0.999 to 0.999, so that the angle would
not be exactly pi. This solved the issue.